WIP types for actions and api
